Fix: 알림 관련 ios 실행 에러
