+++ package.json
@@ -9,11 +9,14 @@
   },
   "dependencies": {
     "@expo/vector-icons": "^13.0.0",
+    "@notifee/react-native": "^7.8.0",
     "@react-native-async-storage/async-storage": "1.18.2",
     "@react-native-community/datetimepicker": "7.2.0",
     "@react-native-community/push-notification-ios": "^1.11.0",
     "@react-native-firebase/app": "^18.2.0",
+    "@react-native-firebase/crashlytics": "^18.3.0",
     "@react-native-firebase/messaging": "^18.2.0",
+    "@react-native-firebase/perf": "^18.3.0",
     "@react-native-picker/picker": "2.4.10",
     "@react-navigation/bottom-tabs": "^6.5.8",
     "@react-navigation/native": "^6.1.7",
@@ -25,8 +28,8 @@
     "expo-device": "~5.4.0",
     "expo-face-detector": "~12.2.0",
     "expo-image-picker": "~14.3.2",
-    "expo-modules-core": "~1.5.8",
-    "expo-splash-screen": "~0.20.4",
+    "expo-modules-core": "~1.5.9",
+    "expo-splash-screen": "~0.20.5",
     "expo-status-bar": "~1.6.0",
     "firebase": "^9.23.0",
     "form-data": "^4.0.0",
@@ -40,7 +43,6 @@
     "react-native-gradle-plugin": "^0.71.19",
     "react-native-modal-datetime-picker": "^17.0.0",
     "react-native-pager-view": "6.2.0",
-    "react-native-push-notification": "^8.1.1",
     "react-native-reanimated": "~3.3.0",
     "react-native-safe-area-context": "4.6.3",
     "react-native-screens": "~3.22.0",
